# ARROW-10540: [Rust] Extended filter kernel to all types and improved performance