--- a/rust/arrow/Cargo.toml
+++ b/rust/arrow/Cargo.toml
@@ -133,3 +133,7 @@ harness = false
 [[bench]]
 name = "concatenate_kernel"
 harness = false
+
+[[bench]]
+name = "mutable_array"
+harness = false