--- a/rust/arrow/src/array/transform/utils.rs
+++ b/rust/arrow/src/array/transform/utils.rs
@@ -72,6 +72,7 @@ pub(super) unsafe fn get_last_offset<T: OffsetSizeTrait>(
     //  Soundness
     //      * offset buffer is always extended in slices of T and aligned accordingly.
     //      * Buffer[0] is initialized with one element, 0, and thus `mutable_offsets.len() - 1` is always valid.
-    let offsets = offset_buffer.data().align_to::<T>().1;
+    let (prefix, offsets, suffix) = offset_buffer.data().align_to::<T>();
+    debug_assert!(prefix.len() == 0 && suffix.len() == 0);
     *offsets.get_unchecked(offsets.len() - 1)
 }