@Dandandan 2020-12-19
I looked a bit at this PR, looks good, didn't find anything weird (but also don't know the details of this part of the code).
Speed ups look great! Any idea where the filter u8 regression comes from?