@jorgecarleitao 2020-12-19
> Any idea where the filter u8 regression comes from?

I tried to explain it in the comment above (there was a race condition here). Does it make sense?