@jorgecarleitao 2020-12-24
Thank you all for your comments and suggestions. Really helpful.

I addressed them all except the `get_last_offset` one, which IMO needs a better approach and thus kept it as is and will work on it later.

This was also rebased against latest master.